+++ weblate/trans/tests/utils.py
@@ -326,6 +326,13 @@ class RepoTestMixin(object):
             'json-mono/en.json',
         )
 
+    def create_joomla(self):
+        return self._create_subproject(
+            'joomla',
+            'joomla/*.ini',
+            'joomla/en-GB.ini',
+        )
+
     def create_tsv(self):
         return self._create_subproject(
             'csv',